+++ s/element/silver.ts
@@ -1,11 +1,11 @@
 
-import {TemplateResult, render} from "lit"
+import {render} from "lit"
 
 import {BaseElement} from "../base/element.js"
 import {MetallicElement} from "./part/metallic.js"
+import {RenderResult} from "../nexus/parts/types.js"
 import {debounce} from "../tools/debounce/debounce.js"
 import {explode_promise} from "../tools/explode_promise.js"
-import { RenderResult } from "../nexus/parts/types.js"
 
 export class SilverElement extends MetallicElement implements BaseElement {
 	#init? = explode_promise<void>()